### 2024-11-29 #675 @njtierney
geotargets: extending targets to produce geospatial pipelines at scale
Submitting Author Name: Nicholas Tierney
Submitting Author Github Handle: @njtierney
Other Package Authors Github handles: @aariq
Repository:  https://github.com/njtierney/geotargets
Version submitted: 0.2.0.9000
Submission type: Standard

Version accepted: TBD

---

```
Package: geotargets
Title: 'Targets' Extensions for Geospatial Formats
Version: 0.2.0.9000
    person(
      given = "Nicholas",
      family = "Tierney",
      email = "nicholas.tierney@gmail.com",
      comment = c(ORCID = "0000-0003-1460-8722")
      ),
    person(
      given = "Eric",
      family = "Scott",
      role = c("aut"),
      comment = c(ORCID = "0000-0002-7430-7879")
    ),
    person(
      given = "Andrew",
      family = "Brown",
      role = c("aut"),
      comment = c(ORCID = "0000-0002-4565-533X")
    )
    )
Description: Provides extensions for various geospatial file formats,
  such as shapefiles and rasters. Currently provides support for the 'terra'
  geospatial formats. See the vignettes for worked examples, demonstrations,
  and explanations of how to use the various package extensions.
Language: en-GB
    targets (>= 1.8.0),
    rlang (>= 1.1.3),
    cli (>= 3.6.2),
    terra (>= 1.7.71),
    withr (>= 3.0.0),
    zip,
    lifecycle
    crew (>= 0.9.2),
    ncmeta,
    sf,
    stars,
    fs
```

The [`geotargets`](https://github.com/njtierney/geotargets) package builds upon the [`targets`](https://github.com/ropensci/targets) package to facilitate geospatial computation, and skip up-to-date operations in geospatial pipelines. [`geotargets`](https://github.com/njtierney/geotargets) means the user doesn't need to write custom targets code to manage the special read/write/wrap/unwrap that is required when using packages like [`terra`](https://github.com/rspatial/terra). It also facilitates dynamic branching, so a raster can be split into parts and then jobs farmed out in parallel and later combined.

The [`geotargets`](https://github.com/njtierney/geotargets) package is for people working with geospatial data with packages like [`terra`](https://github.com/rspatial/terra). Workflows include analysis of sea ice changes over time (https://github.com/njtierney/icebreaker), through to creating statistical pipelines to estimate mosquito insecticide resistance (https://github.com/idem-lab/map-ir-pipeline).

The [`targets`](https://docs.ropensci.org/targets) package does provide workflow automation, but in order for it to work with geospatial  data it requires custom code. The [`geotargets`](https://github.com/njtierney/geotargets) package helps users continue working with [`targets`](https://docs.ropensci.org/targets) with minimal changes to workflow. Other packages like [`chopin`](https://github.com/ropensci/chopin) provide ways to parallelise spatial operations, but this seems distinct from automation, although related.

N/A

N/A

```
── geotargets 0.2.0.9000 ─────────────────────────────────────────────────

✔ Package name is available
✔ has a 'codemeta.json' file.
✔ has a 'contributing' file.
✔ uses 'roxygen2'.
✔ 'DESCRIPTION' has a URL field.
✔ 'DESCRIPTION' has a BugReports field.
✔ Package has at least one HTML vignette
✔ All functions have examples.
✔ Package has continuous integration checks.
✔ Package coverage is 97.7%.
✖ R CMD check found 1 error.
✔ R CMD check found no warnings.
ℹ Function names are duplicated in other packages

ℹ Current status:
✖ This package is not ready to be submitted.
```

I could not find the R CMD Check error, and all checks pass on GitHub Actions - https://github.com/njtierney/geotargets/actions/workflows/R-CMD-check.yaml

We would be interested in submitting this to JOSS, if possible.
