+++ tests/testthat/test-tar-stars.R
@@ -14,7 +14,7 @@ targets::tar_test("tar_stars() works", {
   x <- targets::tar_read(test_stars)
   expect_s3_class(x, "stars")
   expect_snapshot(
-      x
+    x
   )
 })
 
@@ -38,16 +38,19 @@ targets::tar_test("tar_stars_proxy() works", {
 targets::tar_test("tar_stars(mdim=TRUE) works", {
   targets::tar_script({
     geotargets::geotargets_option_set(gdal_raster_driver = "netCDF")
-    list(geotargets::tar_stars(test_stars_mdim, {
-      set.seed(135)
-      m <- matrix(runif(10), 2, 5)
-      names(dim(m)) <- c("stations", "time")
-      times <- as.Date("2022-05-01") + 1:5
-      pts <- sf::st_as_sfc(c("POINT(0 1)", "POINT(3 5)"))
-      s <- stars::st_as_stars(list(Precipitation = m)) |>
-            stars::st_set_dimensions(1, values = pts) |>
-            stars::st_set_dimensions(2, values = times)
-    }, mdim = TRUE))
+    list(geotargets::tar_stars(test_stars_mdim,
+      {
+        set.seed(135)
+        m <- matrix(runif(10), 2, 5)
+        names(dim(m)) <- c("stations", "time")
+        times <- as.Date("2022-05-01") + 1:5
+        pts <- sf::st_as_sfc(c("POINT(0 1)", "POINT(3 5)"))
+        s <- stars::st_as_stars(list(Precipitation = m)) |>
+          stars::st_set_dimensions(1, values = pts) |>
+          stars::st_set_dimensions(2, values = times)
+      },
+      mdim = TRUE
+    ))
   })
 
   targets::tar_make()
@@ -59,44 +62,51 @@ targets::tar_test("tar_stars(mdim=TRUE) works", {
 
 targets::tar_test("tar_stars(mdim=TRUE, ncdf=TRUE) works", {
   targets::tar_script({
-    list(geotargets::tar_stars(test_stars_mdim_ncdf, {
-      set.seed(135)
-      m <- matrix(runif(10), 2, 5)
-      names(dim(m)) <- c("stations", "time")
-      times <- as.Date("2022-05-01") + 1:5
-      pts <- sf::st_as_sfc(c("POINT(0 1)", "POINT(3 5)"))
-      s <- stars::st_as_stars(list(Precipitation = m)) |>
-        stars::st_set_dimensions(1, values = pts) |>
-        stars::st_set_dimensions(2, values = times)
-      s
-    }, driver = "netCDF", mdim = TRUE, ncdf = TRUE))
+    list(geotargets::tar_stars(test_stars_mdim_ncdf,
+      {
+        set.seed(135)
+        m <- matrix(runif(10), 2, 5)
+        names(dim(m)) <- c("stations", "time")
+        times <- as.Date("2022-05-01") + 1:5
+        pts <- sf::st_as_sfc(c("POINT(0 1)", "POINT(3 5)"))
+        s <- stars::st_as_stars(list(Precipitation = m)) |>
+          stars::st_set_dimensions(1, values = pts) |>
+          stars::st_set_dimensions(2, values = times)
+        s
+      },
+      driver = "netCDF",
+      mdim = TRUE,
+      ncdf = TRUE
+    ))
   })
 
   targets::tar_make()
   # warnings related to no CRS
-  suppressWarnings({x <- targets::tar_read(test_stars_mdim_ncdf)})
+  suppressWarnings({
+    x <- targets::tar_read(test_stars_mdim_ncdf)
+  })
   expect_s3_class(x, "stars")
   expect_snapshot(x)
 })
 
[…]
