+++ tests/testthat/test-check-pkg-installed.R
@@ -1,7 +1,7 @@
 test_that("check_pkg_installed() works", {
   expect_snapshot(
-      error = TRUE,
-      x = check_pkg_installed("asdf")
+    error = TRUE,
+    x = check_pkg_installed("asdf")
   )
   expect_null(check_pkg_installed("targets"))
 })
