+++ R/release_bullets.R
@@ -1,6 +1,6 @@
-#see ?usethis::use_release_issue()
+# see ?usethis::use_release_issue()
 release_bullets <- function() {
-    c(
-        "update CITATION.cff with `cffr::cff_write(dependencies = FALSE)` (after incrementing version)"
-    )
+  c(
+    "update CITATION.cff with `cffr::cff_write(dependencies = FALSE)` (after incrementing version)"
+  )
 }

[…]
