### 2024-11-29 5933e95
some small tweaks to improve code coverage and add tests, ignore some other files
