+++ R/tar-terra-rast.R
@@ -88,9 +88,7 @@ tar_terra_rast <- function(name,
   preserve_metadata <- rlang::arg_match0(preserve_metadata, c("drop", "zip"))
 
   # ensure that user-passed `resources` doesn't include `custom_format`
-  if ("custom_format" %in% names(resources)) {
-    cli::cli_abort("{.val custom_format} cannot be supplied to targets created with {.fn tar_terra_rast}")
-  }
+  check_user_resources(resources)
 
   name <- targets::tar_deparse_language(substitute(name))
 
