### 2024-11-29 4b55430
better local tests for check user resources
