+++ R/tar-terra-rast.R
@@ -159,73 +159,74 @@ tar_rast_read <- function(preserve_metadata) {
 }
 
 tar_rast_write <- function(filetype, gdal, preserve_metadata) {
-  switch(
-      preserve_metadata,
-      zip = function(object, path) {
-          #write the raster in a fresh local tempdir() that disappears when function is done
-          tmp <- withr::local_tempdir()
-          tmppath <- file.path(tmp, dirname(path))
-          dir.create(tmppath, recursive = TRUE)
-          terra::writeRaster(
-              object,
-              file.path(tmp, path),
-              filetype = filetype,
-              overwrite = TRUE,
-              gdal = gdal
-          )
-          #package files into a zip file using `zip::zip()`
-          raster_files <- list.files(tmppath, full.names = TRUE)
-          zip::zip(
-              file.path(tmp, basename(path)),
-              files = raster_files,
-              compression_level = 1,
-              mode = "cherry-pick",
-              root = dirname(raster_files)[1]
-          )
+    switch(
+        preserve_metadata,
+        zip = function(object, path) {
+            #write the raster in a fresh local tempdir() that disappears when function is done
+            tmp <- withr::local_tempdir()
+            tmppath <- file.path(tmp, dirname(path))
+            dir.create(tmppath, recursive = TRUE)
+            terra::writeRaster(
+                object,
+                file.path(tmp, path),
+                filetype = filetype,
+                overwrite = TRUE,
+                gdal = gdal
+            )
+            #package files into a zip file using `zip::zip()`
+            raster_files <- list.files(tmppath, full.names = TRUE)
+            zip::zip(
+                file.path(tmp, basename(path)),
+                files = raster_files,
+                compression_level = 1,
+                mode = "cherry-pick",
+                root = dirname(raster_files)[1]
+            )
 
           # move the zip file to the expected place
           file.copy(tmppath, path)
           unlink(tmppath)
-      },
-      gdalraster_sozip = function(object, path) {
-          tmp <- withr::local_tempdir()
-          tmppath <- file.path(tmp, dirname(path))
-          dir.create(tmppath, recursive = TRUE)
+        },
+        gdalraster_sozip = function(object, path) {
 
-          terra::writeRaster(
-              object,
-              file.path(tmp, path),
-              filetype = filetype,
-              overwrite = TRUE,
-              gdal = gdal
-          )
+            tmp <- withr::local_tempdir()
+            tmppath <- file.path(tmp, dirname(path))
+            dir.create(tmppath, recursive = TRUE)
 
-          raster_files <- list.files(tmppath, full.names = TRUE)
+            terra::writeRaster(
+                object,
+                file.path(tmp, path),
+                filetype = filetype,
+                overwrite = TRUE,
+                gdal = gdal
+            )
 
-          # create seek-optimized zip file using gdalraster
-          gdalraster::addFilesInZip(
-              path,
-              raster_files,
-              full_paths = FALSE,
-              overwrite = TRUE,
-              sozip_enabled = "YES",
-              num_threads = 1,
-              quiet = TRUE
-          )
-          # always create sozip regardless of file size (sozip_enabled = "YES")
-          # TODO: allow user control of number of threads?
-          #       how does num_threads interact multiple workers etc.?
+            raster_files <- list.files(tmppath, full.names = TRUE)
 
-          unlink(tmppath)
-      },
-      drop = function(object, path) {
-          terra::writeRaster(
-              object,
-              path,
-              filetype = filetype,
-              overwrite = TRUE,
-              gdal = gdal
-          )
-      }
-  )
+            # create seek-optimized zip file using gdalraster
+            gdalraster::addFilesInZip(
+                path,
+                raster_files,
+                full_paths = FALSE,
[…]
