### 2024-12-02 6f092b8
Revert "variable for temp file path"
This reverts commit f4b7a267ab8b789f6b67044ae70d0876a0ad19bd.
