+++ R/tar-terra-rast.R
@@ -89,7 +89,7 @@ tar_terra_rast <- function(name,
 
   # ensure that user-passed `resources` doesn't include `custom_format`
   check_user_resources(resources)
-
+  
   if (preserve_metadata == "gdalraster_sozip") {
       check_pkg_installed("gdalraster")
   }
@@ -164,8 +164,7 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
         zip = function(object, path) {
             #write the raster in a fresh local tempdir() that disappears when function is done
             tmp <- withr::local_tempdir()
-            tmppath <- file.path(tmp, dirname(path))
-            dir.create(tmppath, recursive = TRUE)
+            dir.create(file.path(tmp, dirname(path)), recursive = TRUE)
             terra::writeRaster(
                 object,
                 file.path(tmp, path),
@@ -174,7 +173,7 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
                 gdal = gdal
             )
             #package files into a zip file using `zip::zip()`
-            raster_files <- list.files(tmppath, full.names = TRUE)
+            raster_files <- list.files(file.path(tmp, dirname(path)), full.names = TRUE)
             zip::zip(
                 file.path(tmp, basename(path)),
                 files = raster_files,
@@ -184,14 +183,14 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
             )
 
           # move the zip file to the expected place
-          file.copy(tmppath, path)
-          unlink(tmppath)
+          file.copy(file.path(tmp, basename(path)), path)
+          unlink(file.path(tmp, basename(path)))
         },
         gdalraster_sozip = function(object, path) {
 
             tmp <- withr::local_tempdir()
-            tmppath <- file.path(tmp, dirname(path))
-            dir.create(tmppath, recursive = TRUE)
+
+            dir.create(file.path(tmp, dirname(path)), recursive = TRUE)
 
             terra::writeRaster(
                 object,
@@ -201,7 +200,7 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
                 gdal = gdal
             )
 
-            raster_files <- list.files(tmppath, full.names = TRUE)
+            raster_files <- list.files(file.path(tmp, dirname(path)), full.names = TRUE)
 
             # create seek-optimized zip file using gdalraster
             gdalraster::addFilesInZip(
@@ -217,7 +216,7 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
             # TODO: allow user control of number of threads?
             #       how does num_threads interact multiple workers etc.?
 
-            unlink(tmppath)
+            unlink(file.path(tmp, path))
         },
         drop = function(object, path) {
             terra::writeRaster(
