### 2024-12-02 dd5210a
temp paths for gdalraster_sozip method only
