+++ R/tar-terra-rast.R
@@ -89,7 +89,7 @@ tar_terra_rast <- function(name,
 
   # ensure that user-passed `resources` doesn't include `custom_format`
   check_user_resources(resources)
-  
+
   if (preserve_metadata == "gdalraster_sozip") {
       check_pkg_installed("gdalraster")
   }
@@ -189,18 +189,19 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
         gdalraster_sozip = function(object, path) {
 
             tmp <- withr::local_tempdir()
-
-            dir.create(file.path(tmp, dirname(path)), recursive = TRUE)
+            dirpath <- file.path(tmp, dirname(path))
+            tmppath <- file.path(tmp, path)
+            dir.create(dirpath, recursive = TRUE)
 
             terra::writeRaster(
                 object,
-                file.path(tmp, path),
+                tmppath,
                 filetype = filetype,
                 overwrite = TRUE,
                 gdal = gdal
             )
 
-            raster_files <- list.files(file.path(tmp, dirname(path)), full.names = TRUE)
+            raster_files <- list.files(dirpath, full.names = TRUE)
 
             # create seek-optimized zip file using gdalraster
             gdalraster::addFilesInZip(
@@ -216,7 +217,7 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
             # TODO: allow user control of number of threads?
             #       how does num_threads interact multiple workers etc.?
 
-            unlink(file.path(tmp, path))
+            unlink(tmppath)
         },
         drop = function(object, path) {
             terra::writeRaster(
