### 2024-12-02 8c4afd7
temp paths for zip method
