### 2024-12-02 9033bce
add test of gdalraster_sozip metadata
