+++ tests/testthat/test-tar-terra-rast.R
@@ -151,3 +151,28 @@ tar_test("metadata is maintained", {
   expect_equal(terra::units(x), rep("m", 3))
   expect_equal(terra::time(x), as.Date("2024-10-01") + c(0, 1, 2))
 })
+
+tar_test("metadata is maintained (gdalraster SOZip)", {
+  skip_if_not_installed("gdalraster")
+  tar_script({
+    library(targets)
+    library(geotargets)
+    library(terra)
+    geotargets_option_set(terra_preserve_metadata = "gdalraster_sozip")
+    make_rast <- function() {
+      f <- system.file("ex/elev.tif", package = "terra")
+      r <- terra::rast(f)
+      r <- c(r, r + 10, r / 2)
+      terra::units(r) <- rep("m", 3)
+      terra::time(r) <- as.Date("2024-10-01") + c(0, 1, 2)
+      r
+    }
+    list(
+      tar_terra_rast(r, make_rast())
+    )
+  })
+  tar_make()
+  x <- tar_read(r)
+  expect_equal(terra::units(x), rep("m", 3))
+  expect_equal(terra::time(x), as.Date("2024-10-01") + c(0, 1, 2))
+})
