### 2024-12-04 #675 @emilyriederer
Thanks for the submission @njtierney ! We're excited to move forward on this review.

I'm happy to introduce @ldecicco-USGS to take it from here as editor.
