+++ R/tar-terra-rast.R
@@ -150,31 +150,31 @@ tar_rast_write <- function(filetype, gdal, preserve_metadata) {
     zip = function(object, path) {
       # write the raster in a fresh local tempdir() that disappears when function is done
       tmp <- withr::local_tempdir()
-      dir.create(file.path(tmp, dirname(path)), recursive = TRUE)
+      raster_tmp_file <- file.path(tmp, basename(path))
+      zip_tmp_file <- file.path(tmp, "object.zip")
       terra::writeRaster(
         object,
-        file.path(tmp, path),
+        filename = raster_tmp_file,
         filetype = filetype,
         overwrite = TRUE,
         gdal = gdal
       )
       # package files into a zip file using `zip::zip()`
-      raster_files <- list.files(file.path(tmp, dirname(path)), full.names = TRUE)
+      raster_files <- list.files(path = tmp, full.names = TRUE)
       zip::zip(
-        file.path(tmp, basename(path)),
+        zipfile = zip_tmp_file,
         files = raster_files,
         compression_level = 1,
         mode = "cherry-pick",
-        root = dirname(raster_files)[1]
+        root = tmp
       )
       # move the zip file to the expected place
-      file.copy(file.path(tmp, basename(path)), path)
-      unlink(file.path(tmp, basename(path)))
+      file.copy(zip_tmp_file, path)
     },
     drop = function(object, path) {
       terra::writeRaster(
         object,
-        path,
+        filename = path,
         filetype = filetype,
         overwrite = TRUE,
         gdal = gdal
