### 2024-12-17 cf512c2
update snapshots (terra internal data changed)
