+++ tests/testthat/_snaps/tile-funs.md
@@ -20,7 +20,7 @@
       
       [[2]]
            xmin      xmax      ymin      ymax 
-       6.133333  6.533333 49.816667 50.191667 
+       6.141667  6.533333 49.816667 50.191667 
       
       [[3]]
            xmin      xmax      ymin      ymax 
@@ -28,6 +28,6 @@
       
       [[4]]
            xmin      xmax      ymin      ymax 
-       6.133333  6.533333 49.441667 49.816667 
+       6.141667  6.533333 49.441667 49.816667 
       
 
