### 2024-12-18 #675 @Aariq
The failing tests are due to updates to the data files included with `terra`.  I'll update these snapshots.
