+++ tests/testthat/_snaps/tar-terra-vect.md
@@ -9,11 +9,11 @@
        extent      : 5.74414, 6.528252, 49.44781, 50.18162  (xmin, xmax, ymin, ymax)
        source      : test_terra_vect
        coord. ref. : lon/lat WGS 84 (EPSG:4326) 
-       names       :  ID_1   NAME_1  ID_2   NAME_2  AREA   POP
-       type        : <num>    <chr> <num>    <chr> <num> <int>
-       values      :     1 Diekirch     1 Clervaux   312 18081
-                         1 Diekirch     2 Diekirch   218 32543
-                         1 Diekirch     3  Redange   259 18664
+       names       :  ID_1   NAME_1  ID_2   NAME_2  AREA       POP
+       type        : <num>    <chr> <num>    <chr> <num>     <num>
+       values      :     1 Diekirch     1 Clervaux   312 1.808e+04
+                         1 Diekirch     2 Diekirch   218 3.254e+04
+                         1 Diekirch     3  Redange   259 1.866e+04
 
 ---
 
@@ -26,9 +26,9 @@
        extent      : 5.74414, 6.528252, 49.44781, 50.18162  (xmin, xmax, ymin, ymax)
        source      : test_terra_vect_shz} (test_terra_vect_shz)
        coord. ref. : lon/lat WGS 84 (EPSG:4326) 
-       names       :  ID_1   NAME_1  ID_2   NAME_2  AREA   POP
-       type        : <num>    <chr> <num>    <chr> <num> <int>
-       values      :     1 Diekirch     1 Clervaux   312 18081
-                         1 Diekirch     2 Diekirch   218 32543
-                         1 Diekirch     3  Redange   259 18664
+       names       :  ID_1   NAME_1  ID_2   NAME_2  AREA       POP
+       type        : <num>    <chr> <num>    <chr> <num>     <num>
+       values      :     1 Diekirch     1 Clervaux   312 1.808e+04
+                         1 Diekirch     2 Diekirch   218 3.254e+04
+                         1 Diekirch     3  Redange   259 1.866e+04
 
