### 2024-12-18 36ecb15
small additions to README:
- add tar_terra_sds/tar_terra_tiles
- add example repos, demo-geotargets, and icebreaker
- remove note on development - it is currently maturing
