### 2024-12-23 #675 @ldecicco-USGS
I expect that with the holidays, we'll slide the due date out a week or so if needed.
