### 2025-01-06 8b683ee
add rOpenSci review badge
