### 2025-01-06 #675 @Aariq
>There is still a problem with some of the SpatRaster attributes not being preserved in in SpatRaster targets using tar_terra_rast() with geotargets::geotargets_option_set(terra_preserve_metadata = "zip"): namely, the varnames and longnames attributes.

This is unfortunately a limitation of `terra`.

``` r
library(terra)
#> terra 1.8.5
r <- terra::rast(system.file("ex/elev.tif", package="terra"))
varnames(r) <- "elev"
longnames(r) <- "elevation above sea level in meters"
r
#> class       : SpatRaster
#> dimensions  : 90, 95, 1  (nrow, ncol, nlyr)
#> resolution  : 0.008333333, 0.008333333  (x, y)
#> extent      : 5.741667, 6.533333, 49.44167, 50.19167  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326)
#> source      : elev.tif
#> name        : elevation
#> min value   :       141
#> max value   :       547

writeRaster(r, "test.tiff", overwrite = TRUE)
r2 <- rast("test.tiff")
varnames(r2)
#> [1] "test"
longnames(r2)
#> [1] ""
```
