### 2025-01-06 #675 @Aariq
## Vignettes

>In the [Raster metadata](http://geotargets.njtierney.com/articles/geotargets.html#raster-metadata) section of the Using terra with geotargets vignette, the return value for the line terra::units(tar_read(r)) of "" is different from what is returned when I run this locally, "m". This code should be reexamined and updated to ensure consistency.

This was purposeful as a segue into the [next section](http://geotargets.njtierney.com/articles/geotargets.html#raster-metadata) which starts with "You may have noticed that the units for the `r` target above have gone missing."  However, I agree that we should reexamine this—it's probably better if the first example is as simple as possible.

>Not all the code in the vignettes runs successfully. Specifically, code that the user should have populated in their _targets.R script files.

I struggled with the decision between `tar_dir(tar_script(<contents of _targets.R))` and just putting the contents of _targets.R in a code block.  I went with the latter for a couple of reasons: 1) it's what other targetopia packages do in their vignettes (e.g. [stantargets](https://docs.ropensci.org/stantargets/articles/introduction.html)), 2) it allows us to have `_targets.R` and `tar_make()` (and other functions that assume the existence of `_targets.R`) in separate code blocks so as to not overwhelm the reader.  I agree that it is usually ideal if copying and pasting code into the console "just works", but I'm not sure that is the way to go for targetopia packages.  We can definitely add some text (or a callout?) making it more clear that that first code chunk needs to be copied and pasted into `_targets.R` or into `tar_script()`—for example, like in the `stantargets` vignette which says "To use it in a [targets](https://docs.ropensci.org/targets/) pipeline, invoke it from the _targets.R script of the project."  What do you think @njtierney?
