### 2025-01-06 3208d4e
update inheritParams documentation from latest version of `targets`
