+++ R/tar_terra_tiles.R
@@ -56,8 +56,7 @@
 #'             tar_terra_tiles(
 #'                 name = rast_split,
 #'                 raster = my_map,
-#'                 ncol = 2,
-#'                 nrow = 2
+#'                 tile_fun = \(x) tile_grid(x, ncol = 2, nrow = 2)
 #'             )
 #'         )
 #'     })
