+++ man/tar_terra_sprc.Rd
@@ -224,8 +224,8 @@ The \code{iteration} argument is unavailable because it is hard-coded to
 \examples{
 if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
   targets::tar_dir({ # tar_dir() runs code from a temporary directory.
-    library(geotargets)
     targets::tar_script({
+      library(geotargets)
       elev_scale <- function(z = 1, projection = "EPSG:4326") {
         terra::project(
           terra::rast(system.file("ex", "elev.tif", package = "terra")) * z,
@@ -245,7 +245,7 @@ if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
       )
     })
     targets::tar_make()
-    x <- targets::tar_read(raster_elevs)
+    targets::tar_read(raster_elevs)
   })
 }
 }
