+++ NEWS.md
@@ -1,6 +1,7 @@
 # geotargets (development version)
 
 * bugfix by @brownag that fixes use of `file.rename()` in `tar_terra_rast(..., preserve_metadata = "zip")`, which does not work when the temporary directory is on a different partition. (#121, PR #122).
+* Fixed examples for `tar_terra_tiles()`, `tar_terra_sds()`, and `tar_terra_sprc()` as reported by @amart90 as part of [rOpenSci review](https://github.com/ropensci/software-review/issues/675)
 
 # geotargets 0.2.0 (29 November 2024)
 
