### 2025-01-06 #675 @Aariq
## Function Documentation

Thanks for the fixes and suggestions @amart90.  I've made fixes to the examples you pointed out and added a details section to `tar_terra_tiles()` in a draft PR ().  I'll wait for @njtierney to be back from holiday to see if he'd like to add details to other functions before merging.
