### 2025-01-09 97af578
tar_terra_vect: pass additional arguments `...` to `terra::writeVector()`
