### 2025-01-09 ac5fe1d
Add `tar_terra_vrt()`
