### 2025-01-10 njtierney/geotargets#139 @njtierney
Improve testing for test-tile-funs and tar_terra_sprc/sds
Based on comments from @amart90 in https://github.com/ropensci/software-review/issues/675

> * There could be additional testing (of course, there could always be more testing), particularly in the tile utility functions. Here are some examples I thought might be useful.
>
>   * In `test-tile-funs.R` add tests for `tile_blocksize` and `tile_grid`. Additional tests could include that the expected number of tiles are being output and the sum of the number of cells is equal to the number of cells in the input raster.
>   * In the test `"recombined tiles are equal to original"`, it could be helpful to repeat the test for all 3 tile helper functions.
>   * For the tests `"tar_terra_sprc/sds() works"`, you have expectations for the class and snapshot. It might also be helpful to have tests to ensure some of the important metadata comes along, such as making sure the layer names, units, time, variables, and other attributes come along when expected. The snapshots for these tests do not include those metadata, so this may be as easy as assigning some of these attributes in the test SPRC/SDS and adding those fields in the snapshot.
