### 2025-01-10 53a2b4b
add more tests
- add checks and tests for integerish for tile_grid and tile_blocksize
- add a first pass test for tile_n returning the number of tiles being output
