### 2025-01-10 #675 @njtierney
Hi @amart90 and @ldecicco-USGS and @Aariq !

Happy new year!

I'm just working through the review from @amart90 now, writing my own notes before reviewing what @Aariq has said. I should have my comments back this coming week.

Thank you very much for such a constructive review, @amart90 - and also to @Aariq for responding so quickly and already resolving some of the issues raised.
