### 2025-01-16 njtierney/geotargets#141 @njtierney
longnames and varnames aren't overwritten 
See

``` r
geotargets::geotargets_option_set(
    gdal_raster_creation_options =
        c("COMPRESS=DEFLATE", "TFW=YES")
)
targets::tar_script({
    elev_scale <- function(z = 1, projection = "EPSG:4326") {
        rast_elev_scale <- terra::project(
            terra::rast(
                system.file(
                    "ex",
                    "elev.tif",
                    package = "terra"
                )
            ) * z,
            projection
        )
        terra::units(rast_elev_scale) <- "m"
        terra::varnames(rast_elev_scale) <- "new-varnames"
        terra::longnames(rast_elev_scale) <- "really-long-new-name"
        terra::time(rast_elev_scale) <- as.Date("2025-01-15")
        rast_elev_scale
    }
    list(
        geotargets::tar_terra_sprc(
            raster_elevs,
            # two rasters, one unaltered, one scaled by factor of 2 and
            # reprojected to interrupted good homolosine
            command = terra::sprc(list(
                elev_scale(1),
                elev_scale(2, "+proj=igh")
            ))
        )
    )
})
targets::tar_make()
#> ▶ dispatched target raster_elevs
#> ● completed target raster_elevs [0.073 seconds, 36.611 kilobytes]
#> ▶ ended pipeline [0.133 seconds]
x <- targets::tar_read(raster_elevs)
x[1]
#> class       : SpatRaster
#> dimensions  : 90, 95, 1  (nrow, ncol, nlyr)
#> resolution  : 0.008333333, 0.008333333  (x, y)
#> extent      : 5.741667, 6.533333, 49.44167, 50.19167  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326)
#> source      : raster_elevs
#> name        : elevation
#> min value   :       141
#> max value   :       547
#> unit        :         m
#> time (days) : 2025-01-15
# works
terra::units(x[1])
#> [1] "m"
# doesn't work
terra::varnames(x[1])
#> [1] "raster_elevs"
# doesn't work
terra::longnames(x[1])
#> [1] ""
# works
terra::time(x[1])
#> [1] "2025-01-15"
```

<summary>

Session info
</summary>

``` r
sessioninfo::session_info()
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 4.4.2 (2024-10-31)
#>  os       macOS Sequoia 15.1
#>  system   aarch64, darwin20
#>  ui       X11
#>  language (EN)
#>  collate  en_US.UTF-8
#>  ctype    en_US.UTF-8
#>  tz       Australia/Melbourne
#>  date     2025-01-16
#>  pandoc   3.2.1 @ /opt/homebrew/bin/ (via rmarkdown)
#>
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version    date (UTC) lib source
#>  backports     1.5.0      2024-05-23 [1] CRAN (R 4.4.0)
#>  base64url     1.4        2018-05-14 [1] CRAN (R 4.4.0)
#>  callr         3.7.6      2024-03-25 [1] CRAN (R 4.4.0)
#>  cli           3.6.3      2024-06-21 [1] CRAN (R 4.4.0)
#>  codetools     0.2-20     2024-03-31 [2] CRAN (R 4.4.2)
#>  data.table    1.16.4     2024-12-06 [1] CRAN (R 4.4.1)
#>  digest        0.6.37     2024-08-19 [1] CRAN (R 4.4.1)
#>  evaluate      1.0.1      2024-10-10 [1] CRAN (R 4.4.1)
#>  fastmap       1.2.0      2024-05-15 [1] CRAN (R 4.4.0)
#>  fs            1.6.5      2024-10-30 [1] CRAN (R 4.4.1)
#>  geotargets    0.1.0.9000 2024-11-20 [1] Github (njtierney/geotargets@ddf163b)
#>  glue          1.8.0      2024-09-30 [1] CRAN (R 4.4.1)
#>  htmltools     0.5.8.1    2024-04-04 [1] CRAN (R 4.4.0)
#>  igraph        2.1.3      2025-01-07 [1] CRAN (R 4.4.2)
#>  knitr         1.49       2024-11-08 [1] CRAN (R 4.4.1)
#>  lifecycle     1.0.4      2023-11-07 [1] CRAN (R 4.4.0)
#>  magrittr      2.0.3      2022-03-30 [1] CRAN (R 4.4.0)
#>  pillar        1.10.1     2025-01-07 [1] CRAN (R 4.4.1)
#>  pkgconfig     2.0.3      2019-09-22 [1] CRAN (R 4.4.0)
#>  processx      3.8.5      2025-01-08 [1] CRAN (R 4.4.1)
#>  ps            1.8.1      2024-10-28 [1] CRAN (R 4.4.1)
#>  R6            2.5.1      2021-08-19 [1] CRAN (R 4.4.0)
#>  Rcpp          1.0.13-1   2024-11-02 [1] CRAN (R 4.4.1)
#>  reprex        2.1.1      2024-07-06 [1] CRAN (R 4.4.0)
#>  rlang         1.1.4      2024-06-04 [1] CRAN (R 4.4.0)
#>  rmarkdown     2.29       2024-11-04 [1] CRAN (R 4.4.1)
#>  rstudioapi    0.17.1     2024-10-22 [1] CRAN (R 4.4.1)
#>  secretbase    1.0.3      2024-10-02 [1] CRAN (R 4.4.1)
#>  sessioninfo   1.2.2      2021-12-06 [1] CRAN (R 4.4.0)
#>  targets       1.9.1      2024-12-04 [1] CRAN (R 4.4.1)
#>  terra         1.8-5      2024-12-12 [1] CRAN (R 4.4.1)
#>  tibble        3.2.1      2023-03-20 [1] CRAN (R 4.4.0)
#>  tidyselect    1.2.1      2024-03-11 [1] CRAN (R 4.4.0)
#>  vctrs         0.6.5      2023-12-01 [1] CRAN (R 4.4.0)
#>  withr         3.0.2      2024-10-28 [1] CRAN (R 4.4.1)
#>  xfun          0.50.5     2025-01-15 [1] Github (yihui/xfun@116d689)
#>  yaml          2.3.10     2024-07-26 [1] CRAN (R 4.4.0)
#>
#>  [1] /Users/nick/Library/R/arm64/4.4/library
#>  [2] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
#>
#> ──────────────────────────────────────────────────────────────────────────────
```
