### 2025-01-16 f23e6bd
add tests for sprc and sds to check that values are carried through
