### 2025-01-16 4a2ec81
various roxygenations
