### 2025-01-16 njtierney/geotargets#141 @Aariq
I think this is part of the known issue in `terra` that varnames and longnames aren't saved when a raster is written and read back in.
