### 2025-01-16 njtierney/geotargets#141 @brownag
This is something I have looked at a little bit-- I believe these labels are intended for specific file types such as netCDF, HDF5 but their handling is not consistent in terra. Note that even `wrap()` does not handle them, let alone `writeRaster()`/`rast()`:

``` r
library(terra)
#> terra 1.8.11

x <- rast(system.file("ex", "elev.tif", package="terra"))

varnames(x) <- "new-varnames"
longnames(x) <- "really-long-new-name"

varnames(x)
#> [1] "new-varnames"
longnames(x)
#> [1] "really-long-new-name"

y <- unwrap(wrap(x))

varnames(y)
#> [1] ""
longnames(y)
#> [1] ""
```

I am working on some PRs for terra to improve metadata writing/round-tripping for more drivers (https://github.com/rspatial/terra/pull/1696#issuecomment-2585888366)

I think that varnames and longnames could be stored in metadata tags or "user tags" for most drivers--either internally within the GDAL data file itself (as units and time are for GeoTIFF), or in a PAM (.aux.xml) file. If I get my changes implemented I will keep var/longnames on the radar and will try to make it so these work the same as date, time and custom user tags.

Thinking about the concept of GDAL subdatasets with drivers other than GTiff and SpatRasterCollection, I have a couple other ideas to perhaps improve `tar_terra_sprc()`... but I'll address that in a different issue/PR
