### 2025-01-16 njtierney/geotargets#141 @njtierney
I think we can leave it open as something to fix once this gets resolved in terra
