### 2025-01-16 9027153
remove varnames and longnames from tests
