### 2025-01-16 njtierney/geotargets#142 @njtierney
update CITATION.cff to include year
As per rOpenSci review
