### 2025-01-16 f8eb78b
update CITATION.cff and also add a GH action to keep the citation file up to date when DESCRIPTION or repo has a new release.
