+++ CITATION.cff
@@ -1,17 +1,18 @@
-# -----------------------------------------------------------
-# CITATION file created with {cffr} R package, v0.5.0
+# --------------------------------------------
+# CITATION file created with {cffr} R package
 # See also: https://docs.ropensci.org/cffr/
-# -----------------------------------------------------------
+# --------------------------------------------
  
 cff-version: 1.2.0
 message: 'To cite package "geotargets" in publications use:'
 type: software
 license: MIT
-title: 'geotargets: Targets extensions for geospatial formats'
-version: 0.0.0.9000
+title: 'geotargets: ''Targets'' Extensions for Geospatial Formats'
+version: 0.2.0.9000
 abstract: Provides extensions for various geospatial file formats, such as shapefiles
-  and rasters. See the vignettes for worked examples and demonstrations and explanations
-  of how to use the various package extensions.
+  and rasters. Currently provides support for the 'terra' geospatial formats. See
+  the vignettes for worked examples, demonstrations, and explanations of how to use
+  the various package extensions.
 authors:
 - family-names: Tierney
   given-names: Nicholas
@@ -24,7 +25,7 @@ authors:
   given-names: Andrew
   orcid: https://orcid.org/0000-0002-4565-533X
 repository-code: https://github.com/njtierney/geotargets
-url: https://njtierney.github.io/geotargets/
+url: http://geotargets.njtierney.com
 contact:
 - family-names: Tierney
   given-names: Nicholas
@@ -43,3 +44,255 @@ keywords:
 - targets
 - vector
 - workflow
+references:
+- type: software
+  title: targets
+  abstract: 'targets: Dynamic Function-Oriented ''Make''-Like Declarative Pipelines'
+  notes: Imports
+  url: https://docs.ropensci.org/targets/
+  repository: https://CRAN.R-project.org/package=targets
+  authors:
+  - family-names: Landau
+    given-names: William Michael
+    email: will.landau.oss@gmail.com
+    orcid: https://orcid.org/0000-0003-1878-3253
+  year: '2025'
+  doi: 10.32614/CRAN.package.targets
+  version: '>= 1.8.0'
+- type: software
+  title: rlang
+  abstract: 'rlang: Functions for Base Types and Core R and ''Tidyverse'' Features'
+  notes: Imports
+  url: https://rlang.r-lib.org
+  repository: https://CRAN.R-project.org/package=rlang
+  authors:
+  - family-names: Henry
+    given-names: Lionel
+    email: lionel@posit.co
+  - family-names: Wickham
+    given-names: Hadley
+    email: hadley@posit.co
+  year: '2025'
+  doi: 10.32614/CRAN.package.rlang
+  version: '>= 1.1.3'
+- type: software
+  title: cli
+  abstract: 'cli: Helpers for Developing Command Line Interfaces'
+  notes: Imports
+  url: https://cli.r-lib.org
+  repository: https://CRAN.R-project.org/package=cli
+  authors:
+  - family-names: Csárdi
+    given-names: Gábor
+    email: csardi.gabor@gmail.com
+  year: '2025'
+  doi: 10.32614/CRAN.package.cli
+  version: '>= 3.6.2'
+- type: software
+  title: terra
+  abstract: 'terra: Spatial Data Analysis'
+  notes: Imports
+  url: https://rspatial.org/
+  repository: https://CRAN.R-project.org/package=terra
+  authors:
+  - family-names: Hijmans
+    given-names: Robert J.
[…]
