### 2025-01-17 510f8bc
update the citation code to include year and also add inst/CITATION file
