+++ CITATION.cff
@@ -26,6 +26,7 @@ authors:
   orcid: https://orcid.org/0000-0002-4565-533X
 repository-code: https://github.com/njtierney/geotargets
 url: http://geotargets.njtierney.com
+date-released: '2024-11-29'
 contact:
 - family-names: Tierney
   given-names: Nicholas
