### 2025-01-17 06b1924
use linter, based off of targets linter
