+++ R/release_bullets.R
@@ -1,6 +1,7 @@
-# see ?usethis::use_release_issue()
+# see ?usethis::use_release_issue() #nolint
 release_bullets <- function() {
   c(
-    "update CITATION.cff with `cffr::cff_write(dependencies = FALSE)` (after incrementing version)"
+    "update CITATION.cff with `cffr::cff_write(dependencies = FALSE)`
+    (after incrementing version)"
   )
 }
