+++ R/tar-terra-sprc.R
@@ -3,9 +3,9 @@
 #' Provides a target format for [terra::SpatRasterCollection] objects,
 #'   which have no restriction in the extent or other geometric parameters.
 #'
-#' @param name Symbol, name of the target. A target
-#'   name must be a valid name for a symbol in R, and it
-#'   must not start with a dot. See [targets::tar_target()] for more information.
+#' @param name Symbol, name of the target. A target name must be a valid name
+#'   for a symbol in R, and it must not start with a dot. See
+#'   [targets::tar_target()] for more information.
 #' @param command R code to run the target.
 #' @param pattern Code to define a dynamic branching pattern for a target. See
 #'   [targets::tar_target()] for more information.
@@ -52,26 +52,28 @@
 #'     x <- targets::tar_read(raster_elevs)
 #'   })
 #' }
-tar_terra_sprc <- function(name,
-                           command,
-                           pattern = NULL,
-                           filetype = geotargets_option_get("gdal.raster.driver"),
-                           gdal = geotargets_option_get("gdal.raster.creation.options"),
-                           ...,
-                           tidy_eval = targets::tar_option_get("tidy_eval"),
-                           packages = targets::tar_option_get("packages"),
-                           library = targets::tar_option_get("library"),
-                           repository = targets::tar_option_get("repository"),
-                           error = targets::tar_option_get("error"),
-                           memory = targets::tar_option_get("memory"),
-                           garbage_collection = targets::tar_option_get("garbage_collection"),
-                           deployment = targets::tar_option_get("deployment"),
-                           priority = targets::tar_option_get("priority"),
-                           resources = targets::tar_option_get("resources"),
-                           storage = targets::tar_option_get("storage"),
-                           retrieval = targets::tar_option_get("retrieval"),
-                           cue = targets::tar_option_get("cue"),
-                           description = targets::tar_option_get("description")) {
+tar_terra_sprc <- function(
+        name,
+        command,
+        pattern = NULL,
+        filetype = geotargets_option_get("gdal.raster.driver"),
+        gdal = geotargets_option_get("gdal.raster.creation.options"),
+        ...,
+        tidy_eval = targets::tar_option_get("tidy_eval"),
+        packages = targets::tar_option_get("packages"),
+        library = targets::tar_option_get("library"),
+        repository = targets::tar_option_get("repository"),
+        error = targets::tar_option_get("error"),
+        memory = targets::tar_option_get("memory"),
+        garbage_collection = targets::tar_option_get("garbage_collection"),
+        deployment = targets::tar_option_get("deployment"),
+        priority = targets::tar_option_get("priority"),
+        resources = targets::tar_option_get("resources"),
+        storage = targets::tar_option_get("storage"),
+        retrieval = targets::tar_option_get("retrieval"),
+        cue = targets::tar_option_get("cue"),
+        description = targets::tar_option_get("description")
+) {
   # ensure that user-passed `resources` doesn't include `custom_format`
     check_user_resources(resources)
 
@@ -127,9 +129,9 @@ tar_terra_sprc <- function(name,
 #' Provides a target format for [terra::SpatRasterDataset] objects,
 #'   which hold sub-datasets, each a `SpatRaster` that can have multiple layers.
 #'
-#' @param name Symbol, name of the target. A target
-#'   name must be a valid name for a symbol in R, and it
-#'   must not start with a dot. See [targets::tar_target()] for more information.
+#' @param name Symbol, name of the target. A target name must be a valid name
+#'   for a symbol in R, and it must not start with a dot. See
+#'   [targets::tar_target()] for more information.
 #' @param command R code to run the target.
[…]
