+++ R/tar_terra_tiles.R
@@ -3,9 +3,9 @@
 #' This target factory is useful when a raster is too large or too high
 #' resolution to work on in-memory. It can instead be split into tiles that can
 #' be iterated over using dynamic branching.
-#' @param name Symbol, name of the target. A target
-#'   name must be a valid name for a symbol in R, and it
-#'   must not start with a dot. See [targets::tar_target()] for more information.
+#' @param name Symbol, name of the target. A target name must be a valid name
+#'   for a symbol in R, and it must not start with a dot. See
+#'   [targets::tar_target()] for more information.
 #' @param raster a `SpatRaster` object to be split into tiles.
 #' @param tile_fun a helper function that returns a list of numeric vectors such
 #'   as [tile_grid()], [tile_n()] or [tile_blocksize] specified in one of the
@@ -227,7 +227,9 @@ tar_terra_tiles_raw <- function(
 #' terra::ext(r2)
 #'
 set_window <- function(raster, window) {
-    raster_out <- c(raster) #forces copying the raster, not just the R object pointing to the same raster in memory
+    # forces copying the raster, not just the R object pointing to the same
+    # raster in memory
+    raster_out <- c(raster)
     terra::window(raster_out) <- window
     raster_out
 }
@@ -294,7 +296,11 @@ set_window <- function(raster, window) {
 #'     tar_terra_tiles(
 #'         name = rast_split_2blocks,
 #'         raster = my_map,
-#'         tile_fun = \(x) tile_blocksize(x, n_blocks_row = 2, n_blocks_col = 1),
+#'         tile_fun = \(x) tile_blocksize(
+#'           x,
+#'           n_blocks_row = 2,
+#'           n_blocks_col = 1
+#'           ),
 #'         description = "Each tile is 2 blocks tall, 1 block wide"
 #'     )
 #'     tar_terra_tiles(
@@ -325,7 +331,7 @@ tile_grid <- function(raster, ncol, nrow) {
     n_tiles <- seq_len(nrow(tile_ext))
     tile_list <- lapply(
         n_tiles,
-        \(i) tile_ext[i,]
+        \(i) tile_ext[i, ]
     )
     tile_list
 }
@@ -336,12 +342,12 @@ tile_blocksize <- function(raster, n_blocks_row = 1, n_blocks_col = 1) {
     tile_ext <-
         terra::getTileExtents(
             raster,
-            terra::fileBlocksize(raster)[1,] * c(n_blocks_row, n_blocks_col)
+            terra::fileBlocksize(raster)[1, ] * c(n_blocks_row, n_blocks_col)
         )
     n_tiles <- seq_len(nrow(tile_ext))
     tile_list <- lapply(
         n_tiles,
-        \(i) tile_ext[i,]
+        \(i) tile_ext[i, ]
     )
     tile_list
 }
@@ -359,20 +365,16 @@ tile_n <- function(raster, n) {
     }
     sq <- sqrt(n)
     sq_round <- floor(sq)
-    quotient <- n/sq_round
+    quotient <- n / sq_round
     is_even <- rlang::is_integerish(quotient)
     is_odd <- !is_even
     if (is_even) {
         nrow <- sq_round
-        ncol <- n/nrow
+        ncol <- n / nrow
     }
     if (is_odd) {
         nrow <- sq_round
         ncol <- ceiling(quotient) #round up
-
-        # #alternatively, only use rows??
-        # nrow <- n
-        # ncol <- 1
     }
 
     cli::cli_inform("creating {nrow} * {ncol} = {nrow*ncol} tile extents\n")
@@ -390,7 +392,7 @@ tile_n <- function(raster, n) {
[…]

[…]
