### 2025-01-17 njtierney/geotargets#146 @njtierney
So I think for the moment we can just keep the default driver as GTiff. This is something that could be a bit tricky to identify how to solve, as there are many possible GDAL drivers - it might be that the best solution is to note that in the documentation some GDAL drivers drop metadata?
