### 2025-01-17 njtierney/geotargets#146 @brownag
The first script works fine for me with "COG" driver using GDAL 3.10 and latest terra 1.8.11:

``` r
library(targets)

tar_script({
    library(targets)
    library(geotargets)
    tar_option_set(packages = "terra")
    geotargets_option_set(gdal_raster_driver = "COG")
    list(
        tar_target(
            tif_file,
            system.file("ex/elev.tif", package = "terra"),
            format = "file"
        ),
        tar_terra_rast(
            r,
            {
                rast <- rast(tif_file)
                units(rast) <- "m"
                rast
            },
            preserve_metadata = "zip"
        )
    )
})

tar_make()
#> ▶ dispatched target tif_file
#> ● completed target tif_file [0.016 seconds, 7.994 kilobytes]
#> ▶ dispatched target r
#> ● completed target r [0.006 seconds, 10.264 kilobytes]
#> ▶ ended pipeline [0.239 seconds]

x <- tar_read(r)

library(terra)
#> terra 1.8.11
gdal(lib = "")
#>     gdal     proj     geos
#> "3.10.0"  "9.5.1" "3.13.0"

describe(sources(x)) |>
    tail() |>
    cat(sep = "\n")
#>   Metadata:
#>     STATISTICS_MAXIMUM=547
#>     STATISTICS_MEAN=-9999
#>     STATISTICS_MINIMUM=141
#>     STATISTICS_STDDEV=-9999
#>     UNIT=m

x
#> class       : SpatRaster
#> dimensions  : 90, 95, 1  (nrow, ncol, nlyr)
#> resolution  : 0.008333333, 0.008333333  (x, y)
#> extent      : 5.741667, 6.533333, 49.44167, 50.19167  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326)
#> source      : r
#> name        : elevation
#> min value   :       141
#> max value   :       547
#> unit        :         m
```
