### 2025-01-17 njtierney/geotargets#146 @brownag
Disregard previous comment, I had my generalizations of metadata reading and writing that I will soon be proposing in terra installed.
