### 2025-01-17 njtierney/geotargets#146 @brownag
Great! Yes, actually I think that was from this fix? https://github.com/rspatial/terra/issues/1688

>  it might be that the best solution is to note that in the documentation some GDAL drivers drop metadata?

I do think we need to have some caveats added to several functions that GDAL version and driver will affect results. However I am hoping a lot of these metadata issues can be resolved with a little elbow grease and focus on making things consistent on the terra side. This should work right through
