### 2025-01-17 njtierney/geotargets#146 @njtierney
> Great! Yes, actually I think that was from this fix? [rspatial/terra#1688](https://github.com/rspatial/terra/issues/1688)
>
> > it might be that the best solution is to note that in the documentation some GDAL drivers drop metadata?
>
> I do think we need to have some caveats added to several functions that GDAL version and driver will affect results. However I am hoping a lot of these metadata issues can be resolved with a little elbow grease and focus on making things consistent on the terra side.

I am really glad that we have this great positive feedback cycle where things we notice here have resulted in improvements in terra - thank you so much for taking the time to identify where these bugs occur, it's really great!

I'll address the documentation in a PR that also provides a minimum version for terra. Overall I think I'll use 1.8-10, as that's what I've currently got, and while it might be inconvenient for users, it is a bit complex to go through and see if this error was fixed at 1.8-7, and we know this works.
