### 2025-01-17 njtierney/geotargets#142 @njtierney
Here's the code I used to add this info and also to add a `CITATION` file at `inst/CITATION`

```r
library(cffr)
cffr::cff_write()
geotargets_cff <- cff_read("CITATION.cff")

new_cff <- geotargets_cff |>
    # date of geotargets v2 release
    # https://github.com/njtierney/geotargets/releases/tag/v0.2.0
    cff_modify("date-released" = "2024-11-29")

cff_validate(new_cff)
cff_write_citation(new_cff, file = "inst/CITATION")
```
