### 2025-01-17 njtierney/geotargets#146 @brownag
Agreed on both things!

1.8-10--makes sense to me to stick to version numbers that correspond to CRAN release/
