### 2025-01-17 d2871d6
Some improvements for metadata being preserved.
- Update to latest version of terra (1.8-10) to ensure metadata is saved for raster drivers like COG.
- Add documentation that details that some versions GDAL and raster drivers might not save metadata and to contact us if there's an issue
- Add test to ensure metadata is maintained for GTiff and COG
