+++ DESCRIPTION
@@ -35,7 +35,7 @@ Imports:
     targets (>= 1.8.0),
     rlang (>= 1.1.3),
     cli (>= 3.6.2),
-    terra (>= 1.7.71),
+    terra (>= 1.8-10),
     withr (>= 3.0.0),
     zip,
     lifecycle
