### 2025-01-17 njtierney/geotargets#148 @njtierney
Changing how targets code is presented in vignettes
From review over in https://github.com/ropensci/software-review/issues/675 - from @amart90:

> * Not all the code in the vignettes runs successfully. Specifically, code that the user should have populated in their _targets.R script files. Examples can be found in [Using `terra` with `geotargets`](http://geotargets.njtierney.com/articles/geotargets.html#tar_terra_rast-targets-with-terra-rasters) and [Example targets pipeline | Dynamic branching with raster tiles](http://geotargets.njtierney.com/articles/tar_terra_tiles.html#example-targets-pipeline). These sections do not not produce a working targets pipeline, so later calls to `tar_make()`, `tar_load()`, etc. throw errors. You call attention that this code should be the contents of the _targets.R file as a comment at the top of the code chunks; however, I think that this is probably too subtle and might trip up even casual targets users. At the very least, I think this deserves a bit more instruction to the user at the first instance in each vignette about copying the code into a _targets.R script file, possibly giving them the suggestion to paste this into `targets::tar_script()`. Another possibility would be to wrap the code in this section in `targets::tar_dir(targets::tar_script({ "code here" }))` the same way it is coded in your examples. This is not the easiest for users to see examples, but it has the benefit of not altering the user's global state (i.e., possibly overwriting an existing _targets.R script file). Once that was addressed, the code in both vignettes ran as expected, with the exception of the case noted above.

I agree with @aariq that we can leave it as is and we will then stay consistent with the rest of the targetopia packages, but I think I really like the idea of the user being able to copy any code chunk and it "just works". However, I think we should add a section at the start of the vignette that discusses how to run these examples - and then we should use `targets::tar_dir(targets::tar_script())` and then give an example of how that code would appear in a `_targets.R` file. I prefer this way because it means users can copy and paste the examples in the vignette easily, and we can refer the user back to that section on how to translate it into a targets pipeline/file.
