### 2025-01-17 njtierney/geotargets#149 @njtierney
Reuse description of problem being solved by tar_terra_* family of targets functions
As noted in the review in https://github.com/ropensci/software-review/issues/675

> * The details section of the `tar_terra_rast` (and `tar_terra_vect`) documentation gives a nice brief explanation of the problem the function is solving and how it addresses it. Other `tar_terra_*` functions (such as `tar_terra_sds`) do not contain a similar explanation, and would benefit from having it, in accordance with the multiple points of enrty principle.

I am now rewriting some of the documentation so it can be reused (https://cran.r-project.org/web/packages/roxygen2/vignettes/reuse.html) in multiple places.
