### 2025-01-17 njtierney/geotargets#151 @njtierney
Test all 3 tile helper functions in test "recombined tiles are equal to original"
From review: https://github.com/ropensci/software-review/issues/675

>   * In the test `"recombined tiles are equal to original"`, it could be helpful to repeat the test for all 3 tile helper functions.
