### 2025-01-17 #675 @njtierney
Firstly, I wanted to thank @amart90 for their feedback and critiques of `geotargets`, they have provided very clear, kind comments, and we really appreciate their review! Thank you also to ldecicco-USGS for being the editor, and to Eric (@aariq) for already responding to several issues and already fixing bugs, and to @brownag for their contributions to terra and geotargets during this review.

I shall respond to the relevant parts inline below - but overall I've opened up a few issues and related pull requests to resolve these. I hope my responses here are clear, and I'll aim to have these addressed in the coming week.

Overall I'm really happy with how the review has improved the package already, and I'm looking forward to seeing what the other reviewers comments are.

> #### Vignettes
> * In the [Raster metadata](http://geotargets.njtierney.com/articles/geotargets.html#raster-metadata) section of the **Using `terra` with `geotargets`** vignette, the return value for the line `terra::units(tar_read(r))` of `""` is different from what is returned when I run this locally, `"m"`. This code should be reexamined and updated to ensure consistency.

Thank you for noting this - I explored this in https://github.com/njtierney/geotargets/issues/146, it seems that if we increase the terra version, we don't get this error. I have written tests for this and improved the documentation at https://github.com/njtierney/geotargets/pull/147

> * Not all the code in the vignettes runs successfully. Specifically, code that the user should have populated in their _targets.R script files. Examples can be found in [Using `terra` with `geotargets`](http://geotargets.njtierney.com/articles/geotargets.html#tar_terra_rast-targets-with-terra-rasters) and [Example targets pipeline | Dynamic branching with raster tiles](http://geotargets.njtierney.com/articles/tar_terra_tiles.html#example-targets-pipeline). These sections do not not produce a working targets pipeline, so later calls to `tar_make()`, `tar_load()`, etc. throw errors. You call attention that this code should be the contents of the _targets.R file as a comment at the top of the code chunks; however, I think that this is probably too subtle and might trip up even casual targets users. At the very least, I think this deserves a bit more instruction to the user at the first instance in each vignette about copying the code into a _targets.R script file, possibly giving them the suggestion to paste this into `targets::tar_script()`. Another possibility would be to wrap the code in this section in `targets::tar_dir(targets::tar_script({ "code here" }))` the same way it is coded in your examples. This is not the easiest for users to see examples, but it has the benefit of not altering the user's global state (i.e., possibly overwriting an existing _targets.R script file). Once that was addressed, the code in both vignettes ran as expected, with the exception of the case noted above.

Thanks for picking up on this! Making the user experience clear and fun/good is important, and often involves a bit of back and forth. I agree with @aariq that we can leave it as is and we will then stay consistent with the rest of the targetopia packages, but I think I really like the idea of the user being able to copy any code chunk and it "just works". However, I think we should add a section at the start of the vignette that discusses how to run these examples - and then we should use `targets::tar_dir(targets::tar_script())` and then give an example of how that code would appear in a `_targets.R` file. I prefer this way because it means users can copy and paste the examples in the vignette easily, and we can refer the user back to that section on how to translate it into a targets pipeline/file.

I've started an issue discussing this at: https://github.com/njtierney/geotargets/issues/148

> #### Function documentation
> * In the documentation for `tar_terra_tiles`, the `tar_terra_tiles` call contains arguments that are inconsistent with the arguments available in this version of geotargets. That is, the example contains arguments for `ncol` and `nrow` but not `tile_fun`. This example should be updated.

Thank you for picking up on this! I note that @aariq has already addressed this in draft PR https://github.com/njtierney/geotargets/pull/136

> * As written, the example code in `tar_terra_sds` and `tar_terra_sprc` does not run successfully - it returns the error `could not find function "tar_terra_sds/tar_terra_sprc"`. These examples should be rewritten. This could be solved by moving `library(geotargets)` within `tar_script()` or by adding the namespace to read `geotargets::tar_terra_sds`. The latter would be consistent with other example code that runs successfully, so that method is prefered.

This is now handled in @aariq's PR: https://github.com/njtierney/geotargets/pull/136

> * In the example of the `tile_grid ` documentation, there is a missing comma after the `)` on the line following `description = "Each tile is 2 blocks tall, 1 block wide"` (line 299 of _R/tar_terra_tiles.R_). This results in an error.

This is now handled in @aariq's PR: https://github.com/njtierney/geotargets/pull/136

> * The details section of the `tar_terra_rast` (and `tar_terra_vect`) documentation gives a nice brief explanation of the problem the function is solving and how it addresses it. Other `tar_terra_*` functions (such as `tar_terra_sds`) do not contain a similar explanation, and would benefit from having it, in accordance with the multiple points of enrty principle.

Good idea! I've started work on this here: https://github.com/njtierney/geotargets/issues/149

> #### Interface
> * It might be helpful to [enumerate possible options](https://design.tidyverse.org/enumerate-options.html) for the geotargets options that have limited possible values. For example, according to the documentation, `terra_preserve_metadata` can be either `"drop"` or `"zip"`; however, I am unconstrained to these options and I am able to set that option to any string. This might be the only option that this comment applies to, as the others are related to external GDAL options. This could handled within the `tar_terra_*()` functions themselves, but it would probably be better implemented in `geotargets_option_set()`. You could enumerate the options in the function parameters and use `rlang::arg_match()` or you could use some other internal logic within the function to ensure the value is not outside the set of possible choices.

I think I generally agree with you here - we decided to use strings to give us multiple possible options, so I think my only concern here is that there might be other options that the user might be able to implement for this?

For the other options, they are GDAL settings and I think in the interest of flexibility, it would be best to leave these open to the user, as there are a lot of GDAL settings, and right now I'm not sure there is a strong benefit to enumerating all of them.

> #### Tests
> * I received a test failure for `"tar_terra_vect() works"` that "`Snapshot of code has changed"`. The cause is that the _POP_ column in the snapshot is class integer whereas the test code produces a SpatVector where the _POP_ column is class numeric. This happened for both `tar_terra_vect()` snapshot tests (lines 29 and 30 of test-tar-terra-vect.R).

I couldn't reproduce this error - and it doesn't look like we can reproduce this on GH actions.

> * There could be additional testing (of course, there could always be more testing), particularly in the tile utility functions. Here are some examples I thought might be useful.
>
>   * In `test-tile-funs.R` add tests for `tile_blocksize` and `tile_grid`. Additional tests could include that the expected number of tiles are being output and the sum of the number of cells is equal to the number of cells in the input raster.

This is a great point - I've added these tests in https://github.com/njtierney/geotargets/pull/140

>   * In the test `"recombined tiles are equal to original"`, it could be helpful to repeat the test for all 3 tile helper functions.

Agreed, I've started an issue here: https://github.com/njtierney/geotargets/issues/151

>   * For the tests `"tar_terra_sprc/sds() works"`, you have expectations for the class and snapshot. It might also be helpful to have tests to ensure some of the important metadata comes along, such as making sure the layer names, units, time, variables, and other attributes come along when expected. The snapshots for these tests do not include those metadata, so this may be as easy as assigning some of these attributes in the test SPRC/SDS and adding those fields in the snapshot.

Agreed, however it appears that this metadata such as units, and time aren't included in the print method for sprc - in fact this information is only printed for each layer. I've added some tests now to ensure they are still provided - however some features such as varname and longnames aren't saved - this is an issue external to geotargets, see https://github.com/njtierney/geotargets/issues/141

>
> #### Functionality
> * There is still a problem with some of the SpatRaster attributes not being preserved in in SpatRaster targets using `tar_terra_rast()` with `geotargets::geotargets_option_set(terra_preserve_metadata = "zip")`: namely, the varnames and longnames attributes.
>
> ```r
>         targets::tar_dir({
>     targets::tar_script({
>         geotargets::geotargets_option_set(terra_preserve_metadata = "zip")
>         modify_rast <- function(r) {
>             r <- c(r, r + 2, r * 2)
>             names(r) <- c("elev", "elev_p_2", "elev_t_2")
>             terra::units(r) <- "m"
>             terra::time(r) <- as.Date(c("2020-01-01", "2020-01-02", "2020-01-03"))
>             terra::varnames(r) <- "elev"
>             terra::longnames(r) <- "elevation above sea level in meters"
>             r
>         }
>
>         list(
>             geotargets::tar_terra_rast(
>                 elev_rast,
>                 terra::rast(system.file("ex/elev.tif", package="terra"))
>             ),
>             geotargets::tar_terra_rast(
>                 modified_rast,
>                 modify_rast(elev_rast)
>             )
>         )
>     })
>
>     targets::tar_make()
>     modified_rast <- targets::tar_read(modified_rast)
>
> })
> ```
>
> I tried using different gdal raster drivers and none of them preserved the varnames or longnames attributes. To me this is not a reason to stall the review, but I do think that a this should be logged as an issue to eventually resolve and there should be a note in the documentation explaining to that these attributes are not preserved.

This is a limitation of terra not geotargets - see discussion in https://github.com/njtierney/geotargets/issues/141

> #### Other comments
> * Consider adding a year to the CITATION.cff file. Currently, `citation("geotargets")` returns a warning that it "`could not determine year for ‘geotargets’ from package DESCRIPTION file`". It could also be useful to add a date (either dateModified or datePublished?) to codemeta.json.

This has been updated and fixed in https://github.com/njtierney/geotargets/issues/142 and https://github.com/njtierney/geotargets/pull/143

> * Two of the badges are currently displaying unexpected/unhelpful information:
>
>   * codecov: unknown. I am not familiar with how the CI for this works, but the [link from the badge](https://app.codecov.io/gh/njtierney/geotargets?branch=master) seems to show a 97.57% code coverage for the most recent commit (855a5e3), so I'm not sure why the badge shows "unknown".
>     -pkgcheck: failing. I'm not very familiar with pkgcheck, but looking at the [logs](https://github.com/njtierney/geotargets/actions/runs/12403450237) it seems like this is a problem with the CI and not a problem with the package contents.

It's unclear to me why we are getting this issue with codecov - I've tried to regenerate the token, I'll see if that helps.

> * Per [1.6 README | rOpenSci Packaging Guide](https://devguide.ropensci.org/pkg_building.html#readme) that the users should be directed to citation information in the README. While GitHub provides a nice "Cite this repository" popup on the right-hand menu that is generated from the CITATION.cff file, there is nothing in the README.

Thanks for this, updates are in https://github.com/njtierney/geotargets/pull/143

> * The development status in codemeta.json (`"https://www.repostatus.org/#wip"`) does not match the development status in the readme (`"www.repostatus.org/#active"`). I think this discrepancy is carried forward to [Targetopia](https://wlandau.github.io/targetopia/packages.html), which shows the status as "WIP".

Thanks for this, updates are in https://github.com/njtierney/geotargets/pull/143

> * There were about 70 instances of lines being greater than 80 characters in width. Most of these were cases where shortening the lines would have made the code more difficult to read (i.e., in code in the examples or in the default parameters in function definitions), but there were cases where the lines could be reduced in width. These cases were mostly just beyond 80 characters in function documentation or in-line comments. While not a requirement, it may be worth using lintr to locate these and shorten as many of these lines as possible.

Thanks for this, we've now added lintr (https://github.com/njtierney/geotargets/pull/143) to help resolve this.
