### 2025-01-17 njtierney/geotargets#141 @Aariq
Hopefully there will be nothing for us to do if these metadata end up getting written with the same mechanisms as layer names or units.
