### 2025-01-17 njtierney/geotargets#148 @Aariq
Sounds good to me. This will also greatly simplify the vignettes by eliminating a lot of duplicate code.
