### 2025-01-17 njtierney/geotargets#141 @brownag
https://search.r-project.org/CRAN/refmans/terra/html/varnames.html

I was reading the documentation for varnames and longnames--and it seems there is some discussion of this issue:

> Each SpatRaster _data source_ can also have a variable name and a long variable name. They are set when reading a file with possibly multiple sub-datasets (e.g. netcdf or hdf5 format) into a single SpatRaster. Each sub-datset is a seperate "data-source" in the SpatRaster. Note that newly created or derived SpatRasters always have a single variable (data source), and therefore the variable names are lost when processing a multi-variable SpatRaster. Thus the variable names are mostly useful to understand a SpatRaster created from some files and for managing SpatRasterDatasets.

It seems "expected" that these names are only read and not written, and are targeted to specific data source types..

However I do think it is a bit confusing that there are exported setters varnames<- and longnames<- if common ops like `wrap()` and `writeRaster()` will wipe them out. I believe there is a strong argument for trying to get them into the PackedSpatRaster and regular metadata so that things like writing to temp files and serialization work properly, not only for our use cases here in geotatrgets. It may be that they can only be supported in a standard way for a subset of drivers, as I do not have band-level user tags working for all drivers. I am still investigating what the best possible implementation is. Worst case in my view--if we can't do it in a standard way via .aux.xml--it would be a few more elements that could be added to the terra-specific aux.json file.
