### 2025-01-20 #675 @njtierney
Hi @ldecicco-USGS - just thought I'd check actually, is it beneficial to wait for all reviewer feedback before addressing changes? I think we are nearly done with the review from @amart90 - so I might wrap that up soon, but happy to wait for other reviews to come back if that makes things easier.
