### 2025-01-20 9c914a3
expand documentation around retaining metadata, and setting units
