### 2025-01-20 d3700c9
minor tweaks to details for spatraster/vect details for tar_terra_*
