### 2025-01-20 be38c14
change pkgcheck to not run on every single push, just once a month is OK with me
