+++ .github/workflows/pkgcheck.yaml
@@ -6,13 +6,9 @@ concurrency:
   cancel-in-progress: true
 
 on:
-  # Manually trigger the Action under Actions/pkgcheck
-  workflow_dispatch:
-  # Run on every push to main
-  push:
-    branches:
-      - main
-      - master
+  schedule:
+    # Runs at midnight UTC on the 1st day of every month
+    - cron: '0 0 1 * *'
 
 jobs:
   pkgcheck:
