### 2025-01-20 #675 @ldecicco-USGS
I vote that you can go ahead and push up your changes. I heard back from @lidefi87 that there was a technical issue installing the rOpenSci pkgstats package. So I don't think pushing up your changes will change her review - and then she doesn't need to provide the same feedback to the areas already addressed.
