### 2025-01-22 9b86da6
use flint for more automated linting
