### 2025-01-23 04015c6
improve test  - use driver that stores metadata externally  - include dataset and band user tags
