### 2025-01-27 #675 @lidefi87
- Mostly yes, but I am suggesting some changes to the README file under *"Review Comments"* below.
  - Some of the vignettes do not run successfully locally. I am including a list in the *"Review Comments"* section below.
  - Some of the examples do not run successfully locally as I mentioned above. See the *"Review Comments"* section for more details.
  - Some information is missing. Details are in the *"Review Comments"* section below.

  - I could not run the test because I was not able to install some of the dependencies of the `pkgcheck` package.
  - Overall the package looks really good and I can see myself using it in my work. However, I think that currently the documentation assumes users are already familiar with the `targets` package. I don't think I have a good suggestion about how to solve this issue and make it more beginner friendly. I feel like most of the context needed is in the `targets` package documentation, but I don't know how to make this clearer. Maybe just suggesting potential users to become familiar with `targets` will suffice, but I am not sure if someone else may have a better suggestion.

---

Thanks again for the invite to review the `geotargets` package and apologies once again for the delay. I have not been having a good time with technology lately.

I have gone through the package and I have the following comments:
- **A statement of need** - Although the README file does state the problem `geotargets` is designed to solve, it assumes that the user is already familiar with the `targets` package. I know there is a link to the `targets` package but I think it would be better if the README file had a brief description of the `targets` package and how `geotargets` builds upon it. I would suggest to copy the first paragraph of the `targets` README file to provide more context. However, if you expect the audience for this package to be `targets` users, then this is not necessary.
- **Vignettes/Examples** - The examples for the following functions did not run without making changes first. Granted, I could have read that the examples required an environmental variable to be set, but I think it would be better if this information was specified just before the `Run examples` link. The functions that did not work without changes are:
  - `geotargets_option_set`, `geotargets_option_get`
  - `tar_stars`, `tar_stars_proxy`
  - `tar_terra_rast`
  - `tar_terra_sds`
  - `tar_terra_sprc`
  - `tar_terra_tiles`
  - `tar_terra_vect`

Once I set the `TAR_LONG_EXAMPLES` environmental variable to `true`, the examples run without issue, but only limited information is returned (mostly messages that the code has been evaluated). Perhaps because I am not very familiar with the `targets` package, I am not sure if this is the expected behavior. From the perspective of a brand new user, I think it would be great if more information could be added about what is expected from these examples as it is not clear to me.

I later realised that I could run the examples for the functions below without using the `tar_dir` function. This allowed me to see what the expected output looked like and even plot it. I think it would be great if the examples for the functions below could be updated to include this information.
  - `tar_stars`
  - `tar_terra_rast`
  - `tar_terra_sds`
  - `tar_terra_sprc`
  - `tar_terra_vect`

I find that being able to see the expected output is really helpful to me, but maybe this is not the case for everyone. I wonder if this has to do with the fact that I am not very familiar with the `targets` package.

- **Community guidelines** - The README file does include this information, but there are no contact details for reporting of any issues. See **Enforcement** in the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/1/CODE_OF_CONDUCT.html).
