### 2025-02-27 b8e88d4
address concerns from reviewer 2 in https://github.com/ropensci/software-review/issues/675
