+++ R/tar-terra-vect.R
@@ -28,6 +28,8 @@
 #'   which both appear to work generally.
 #' @export
 #' @examples
+#' # For CRAN. Ensures these examples run under certain conditions.
+#' # To run this locally, run the code inside this if statement
 #' if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
 #'   targets::tar_dir({ # tar_dir() runs code from a temporary directory.
 #'     targets::tar_script({
