+++ R/tar-stars.R
@@ -38,6 +38,8 @@
 #' @seealso [targets::tar_target()]
 #' @export
 #' @examplesIf rlang::is_installed("stars")
+#' # For CRAN. Ensures these examples run under certain conditions.
+#' # To run this locally, run the code inside this if statement
 #' if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
 #'   targets::tar_dir({ # tar_dir() runs code from a temporary directory.
 #'     library(geotargets)
