+++ R/tar-terra-rast.R
@@ -51,6 +51,8 @@
 #' @seealso [targets::tar_target()]
 #' @export
 #' @examples
+#' # For CRAN. Ensures these examples run under certain conditions.
+#' # To run this locally, run the code inside this if statement
 #' if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
 #'   targets::tar_dir({ # tar_dir() runs code from a temporary directory.
 #'     library(geotargets)
