+++ man/tar_stars.Rd
@@ -269,6 +269,8 @@ The \code{iteration} argument is unavailable because it is hard-coded to
 }
 \examples{
 \dontshow{if (rlang::is_installed("stars")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+# For CRAN. Ensures these examples run under certain conditions.
+# To run this locally, run the code inside this if statement
 if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
   targets::tar_dir({ # tar_dir() runs code from a temporary directory.
     library(geotargets)
