### 2025-02-27 #675 @njtierney
Thank you very much to @lidefi87 for reviewing `geotargets` - we really appreciate you taking the time, and your perspective as a new targets user!

> > #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [x]  **Functionality:** Any functional claims of the software have been confirmed.
> * [x]  **Performance:** Any performance claims of the software have been confirmed.*
> * [ ]  **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
>     * I could not run the test because I was not able to install some of the dependencies of the `pkgcheck` package.* [ ]
>     * **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.
>
>     * Overall the package looks really good and I can see myself using it in my work. However, I think that currently the documentation assumes users are already familiar with the `targets` package. I don't think I have a good suggestion about how to solve this issue and make it more beginner friendly. I feel like most of the context needed is in the `targets` package documentation, but I don't know how to make this clearer. Maybe just suggesting potential users to become familiar with `targets` will suffice, but I am not sure if someone else may have a better suggestion.

The intended audience of `geotargets` is those who know about targets already - we have now added a link to "targets in 4 minutes" in the README.

> ### Review Comments
> Thanks again for the invite to review the `geotargets` package and apologies once again for the delay and thanks for your understanding. I have not been having a good time with technology lately.
>
> I have gone through the package and I have the following comments: **Documentation**
>
> * **A statement of need** - Although the README file does state the problem `geotargets` is designed to solve, it assumes that the user is already familiar with the `targets` package. I know there is a link to the `targets` package but I think it would be better if the README file had a brief description of the `targets` package and how `geotargets` builds upon it. I would suggest to copy the first paragraph of the `targets` README file to provide more context. However, if you expect the audience for this package to be `targets` users, then this is not necessary.

We do expect `geotargets` users to be targets users, as this builds upon targets.

> * **Vignettes/Examples** - The examples for the following functions did not run without making changes first. Granted, I could have read that the examples required an environmental variable to be set, but I think it would be better if this information was specified just before the `Run examples` link. The functions that did not work without changes are:
>
>   * `geotargets_option_set`, `geotargets_option_get`
>   * `tar_stars`, `tar_stars_proxy`
>   * `tar_terra_rast`
>   * `tar_terra_sds`
>   * `tar_terra_sprc`
>   * `tar_terra_tiles`
>   * `tar_terra_vect`
>
> Once I set the `TAR_LONG_EXAMPLES` environmental variable to `true`, the examples run without issue, but only limited information is returned (mostly messages that the code has been evaluated). Perhaps because I am not very familiar with the `targets` package, I am not sure if this is the expected behavior. From the perspective of a brand new user, I think it would be great if more information could be added about what is expected from these examples as it is not clear to me.

I can appreciate that these examples could have more context for users not familiar with targets. Things like `#' if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {` are there to ensure these examples are only run in certain situations, which is to do with CRAN testing. We can add some more comments around these examples to explain what these code examples are doing.

>
> I later realised that I could run the examples for the functions below without using the `tar_dir` function. This allowed me to see what the expected output looked like and even plot it. I think it would be great if the examples for the functions below could be updated to include this information.
>
> * `tar_stars`
> * `tar_terra_rast`
> * `tar_terra_sds`
> * `tar_terra_sprc`
> * `tar_terra_vect`
>
> I find that being able to see the expected output is really helpful to me, but maybe this is not the case for everyone. I wonder if this has to do with the fact that I am not very familiar with the `targets` package.

I agree it is helpful to see the expected output, however, it is a bit of a tricky thing to demonstrate to users, as `tar_script` will write a `_targets.R` file that then gets used by `tar_make()`. I have added a comment "`#' # For CRAN. Ensures these examples run under certain conditions. To run this locally, run the code inside this if statement`" .

>
> * **Community guidelines** - The README file does include this information, but there are no contact details for reporting of any issues. See **Enforcement** in the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/1/CODE_OF_CONDUCT.html).

Thank you for this comment - we have updated this to use the ropensci code of conduct, which is what targets links to: https://ropensci.org/code-of-conduct/
