+++ .github/workflows/flir.yaml
@@ -9,14 +9,14 @@ on:
     types: [published]
   workflow_dispatch:
 
-name: flint
+name: flir
 
 jobs:
-  flint:
+  flir:
     runs-on: macOS-latest
     # Only restrict concurrency for non-PR jobs
     concurrency:
-      group: flint-${{ github.event_name != 'pull_request' || github.run_id }}
+      group: flir-${{ github.event_name != 'pull_request' || github.run_id }}
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     permissions:
@@ -26,10 +26,12 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
 
-      - name: Install flint
-        run: install.packages("flint", repos = c("https://etiennebacher.r-universe.dev/", getOption("repos")))
+      - name: Install flir
+        run: install.packages("flir", repos = c("https://etiennebacher.r-universe.dev/", getOption("repos")))
         shell: Rscript {0}
 
-      - name: Run flint
-        run: flint::lint()
+      - name: Run flir
+        run: flir::lint()
         shell: Rscript {0}
+        env:
+          FLIR_ERROR_ON_LINT: true
