+++ tests/testthat/test-tar-terra-vect.R
@@ -3,9 +3,7 @@ targets::tar_test("tar_terra_vect() works", {
   targets::tar_script({
     lux_area <- function(projection = "EPSG:4326") {
       terra::project(
-        terra::vect(system.file("ex", "lux.shp",
-          package = "terra"
-        )),
+        terra::vect(system.file("ex", "lux.shp", package = "terra")),
         projection
       )
     }
@@ -54,12 +52,11 @@ targets::tar_test("tar_terra_vect() works with dynamic branching", {
   expect_length(targets::tar_read(my_vect_subs), 2)
 })
 
-targets::tar_test(
-    "tar_terra_vect() works with multiple workers (tests un/marshaling)", {
+targets::tar_test("tar_terra_vect() works with multiple workers (tests un/marshaling)", {
   targets::tar_script({
     targets::tar_option_set(
-        controller = crew::crew_controller_local(workers = 2)
-        )
+      controller = crew::crew_controller_local(workers = 2)
+    )
     list(
       geotargets::tar_terra_vect(
         vect1,
