+++ tests/testthat/test-geotargets-option.R
@@ -1,5 +1,4 @@
-targets::tar_test(
-    "geotargets_options_get() retrieves options in correct priority", {
+targets::tar_test("geotargets_options_get() retrieves options in correct priority", {
   # options takes precedent over env var
   withr::with_envvar(
     c("GEOTARGETS_GDAL_RASTER_DRIVER" = "COG"),
@@ -13,7 +12,10 @@ targets::tar_test(
         )
       })
       targets::tar_make()
-      expect_identical(geotargets::geotargets_option_get("gdal.raster.driver"), "GIF")
+      expect_identical(
+        geotargets::geotargets_option_get("gdal.raster.driver"),
+        "GIF"
+      )
     })
   )
 })
@@ -29,8 +31,7 @@ test_that("geotargets_option_set() works", {
   expect_identical(geotargets_option_get("gdal_raster_driver"), "COG")
 })
 
-test_that(
-    "options aren't reset with multiple calls to geotargets_option_set()", {
+test_that("options aren't reset with multiple calls to geotargets_option_set()", {
   op_rast <- getOption("geotargets.gdal.raster.driver")
   withr::defer(options("geotargets.gdal.raster.driver" = op_rast))
   op_vect <- getOption("geotargets.gdal.vector.driver")

[…]
