### 2025-02-27 njtierney/geotargets#141 @amart90
Could some of these attributes (longnames, varnames, time, etc.) that are not preserved when written be written to R objects within the zip file in `tar_rast_write()`? Then when read in with `tar_rast_read()` they could be re-assigned to the SpatRaster? Here is some pseudocode (it's real code, I guess, but I haven't tested it - it's more a demonstration of the concept):

```r
tar_rast_write <- function(filetype, gdal, preserve_metadata) {
  switch(preserve_metadata,
         zip = function(object, path) {
           # write the raster in a fresh local tempdir() that disappears when
           # function is done
           tmp <- withr::local_tempdir()
           raster_tmp_file <- file.path(tmp, basename(path))
           attr_tmp_file <- file.path(tmp, "attrs.RDS")
           rast_attrs <- list(
             time = terra::time(object),
             longnames = terra::longnames(object),
             varnames = terra::varnames(object)
           )
           zip_tmp_file <- file.path(tmp, "object.zip")
           terra::writeRaster(
             object,
             filename = raster_tmp_file,
             filetype = filetype,
             overwrite = TRUE,
             gdal = gdal
           )
           saveRDS(rast_attrs, attr_tmp_file, compress = FALSE)

           # package files into a zip file using `zip::zip()`
           raster_files <- list.files(path = tmp, full.names = TRUE)
           zip::zip(
             zipfile = zip_tmp_file,
             files = raster_files,
             compression_level = 1,
             mode = "cherry-pick",
             root = tmp
           )
           # move the zip file to the expected place
           file.copy(zip_tmp_file, path)
         }
      # ...
  )
}

tar_rast_read <- function(preserve_metadata) {
  switch(preserve_metadata,
         zip = function(path) {
           tmp <- tempdir()
           # NOTE: cannot use withr::local_tempdir() because the unzipped files need
           # to persist so that the resulting `SpatRaster` object doesn't have a
           # broken file pointer
           zip::unzip(zipfile = path, exdir = tmp)
           rast_attrs <- readRDS("attrs.RDS")
           rast_out <- terra::rast(file.path(tmp, basename(path)))
           terra::longnames(rast_out) <- rast_attrs[["longnames"]]
           terra::varnames(rast_out) <- rast_attrs[["varnames"]]
           terra::time(rast_out) <- rast_attrs[["time"]]
           return(rast_out)
         },
         drop = function(path) terra::rast(path)
  )
}
```
