### 2025-02-27 njtierney/geotargets#141 @Aariq
I'm still not sure we should be trying to do anything that `terra` doesn't already do when writing and reading `SpatRaster`s.
