### 2025-02-27 njtierney/geotargets#141 @brownag
I requested this be fixed for `wrap()` and `unwrap()` in https://github.com/rspatial/terra/issues/1719 which should resolve any issues where varnames/longnames would be wiped out in the marshalling proces.

Likewise, I am not sure we should be coming up with our own approach for this. The ability to store these attributes is driver-specific. Now in terra metadata are handled by the GDAL driver (after https://github.com/rspatial/terra/pull/1714), so no terra-specific JSON metadata files anymore. Either attributes are stored internal to the file, or they are put in a Persistent Auxiliary Metadata file (PAM).

Some GDAL drivers support user-defined metadata tags--see `terra::metags()`. If preserving longnames and varnames in a more agnostic way is a desire for us (personally I never use them) it possibly should be done through the metadata user tag mechanism prior to writing a SpatRaster target, and then use an appropriate driver (e.g. GeoTIFF).

It might be reasonable to request storing varnames and longnames in user tags as a feature of terra, but I have not had a chance to investigate all the details of the current implementation and how it would interact with e.g. HDF and netCDF drivers that these attributes were originally intended for.
