### 2025-02-27 njtierney/geotargets#141 @Aariq
The unfortunate thing about this being supported by `wrap()` and `unwrap()` but not by `writeRaster()` and the `aux.json` file is that there may be inconsistent behavior depending on whether an upstream target is skipped or run, whether `memory = "transient"` or `memory = "persistent"` and whether a `crew` controller is used or not.  It would still be best to not use longnames and varnames in a `targets` pipeline.
