### 2025-02-27 njtierney/geotargets#141 @Aariq
Hold up—I just read https://github.com/rspatial/terra/pull/1714.  Does this mean if we require this version of `terra` as minimum we wouldn't need the `preserve_metadata` argument to `tar_terra_rast()` anymore? (sorry, this is a bit off topic)
