### 2025-02-27 njtierney/geotargets#141 @brownag
> The unfortunate thing about this being supported by `wrap()` and `unwrap()` but not by `writeRaster()` and the `aux.json` file is that there may be inconsistent behavior depending on whether an upstream target is skipped or run, whether `memory = "transient"` or `memory = "persistent"` and whether a `crew` controller is used or not. It would still be best to not use longnames and varnames in a `targets` pipeline.

This, I think, is an argument for using something like `terra::wrapCache()` like #106. Alternately, perhaps we could issue messages that at least indicate these specific elements will not be preserved and automatically null them out when running a target

I am not sure it is a good general assumption that all data required to reproduce an R object can be stored in a regular spatial format--even with (auxiliary) metadata support--without also making some efforts to serialize the actual (Packed)SpatRaster/Vector object that was created.

> Hold up—I just read [rspatial/terra#1714](https://github.com/rspatial/terra/pull/1714). Does this mean if we require this version of `terra` as minimum we wouldn't need the `preserve_metadata` argument to `tar_terra_rast()` anymore? (sorry, this is a bit off topic)

Depends on the driver and version of GDAL. For GeoTIFF at least it all can be stored internally, as well as for several other drivers.
