### 2025-03-03 #675 @ldecicco-USGS
@njtierney sounds great!

@lidefi87 and @amart90 , please let us know if you request further changes. If you are satisfied please indicate your approval using [this template](https://devguide.ropensci.org/approval2template.html):

```
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing:
```
