### 2025-03-04 #675 @njtierney
Thanks @ldecicco-USGS !

Quick question - do we need to do anything special for submitting to JOSS, or do we discuss that later?
