### 2025-03-11 #675 @ropensci-review-bot
Approved! Thanks @njtierney for submitting and @amart90, @lidefi87 for your reviews! :grin:
