+++ R/geotargets-option.R
@@ -1,6 +1,6 @@
 #' Get or Set geotargets Options
 #'
-#' Get or set behavior for geospatial data target stores using
+#' Get or set behaviour for geospatial data target stores using
 #' geotargets-specific global options.
 #'
 #' @param gdal_raster_driver character, length 1; set the driver used for raster
@@ -33,7 +33,7 @@
 #'   `preserve_metadata`.
 #'
 #' @details
-#' # Potential issues retaining metdatadata
+#' # Potential issues retaining metadata
 #'
 #'   If you have an issue with retaining metadata (such as units, time, etc),
 #'   this could be due to the versions of GDAL and terra on your machine. We
@@ -86,27 +86,29 @@
 #'   })
 #' }
 #'
-geotargets_option_set <- function(gdal_raster_driver = NULL,
-                                  gdal_raster_creation_options = NULL,
-                                  gdal_vector_driver = NULL,
-                                  gdal_vector_creation_options = NULL,
-                                  terra_preserve_metadata = NULL) {
-    # TODO do this programmatically with formals() or something?
-    # `options()` also accepts a named list
-    gdal_raster_driver <- gdal_raster_driver %||%
-        geotargets_option_get("gdal.raster.driver")
+geotargets_option_set <- function(
+  gdal_raster_driver = NULL,
+  gdal_raster_creation_options = NULL,
+  gdal_vector_driver = NULL,
+  gdal_vector_creation_options = NULL,
+  terra_preserve_metadata = NULL
+) {
+  # TODO do this programmatically with formals() or something?
+  # `options()` also accepts a named list
+  gdal_raster_driver <- gdal_raster_driver %||%
+    geotargets_option_get("gdal.raster.driver")
 
-    gdal_raster_creation_options <- gdal_raster_creation_options %||%
-        geotargets_option_get("gdal.raster.creation.options")
+  gdal_raster_creation_options <- gdal_raster_creation_options %||%
+    geotargets_option_get("gdal.raster.creation.options")
 
-    gdal_vector_driver <- gdal_vector_driver %||%
-        geotargets_option_get("gdal.vector.driver")
+  gdal_vector_driver <- gdal_vector_driver %||%
+    geotargets_option_get("gdal.vector.driver")
 
-    gdal_vector_creation_options <- gdal_vector_creation_options %||%
-        geotargets_option_get("gdal.vector.creation.options")
+  gdal_vector_creation_options <- gdal_vector_creation_options %||%
+    geotargets_option_get("gdal.vector.creation.options")
 
-    terra_preserve_metadata <- terra_preserve_metadata %||%
-        geotargets_option_get("terra.preserve.metadata")
+  terra_preserve_metadata <- terra_preserve_metadata %||%
+    geotargets_option_get("terra.preserve.metadata")
 
   options(
     "geotargets.gdal.raster.driver" = gdal_raster_driver,
@@ -128,13 +130,16 @@ geotargets_option_get <- function(name) {
   option_name <- geotargets_repair_option_name(name)
   # check if `name` is one of the possible options
   option_name <-
-    rlang::arg_match0(option_name, c(
-      "geotargets.gdal.raster.driver",
-      "geotargets.gdal.raster.creation.options",
-      "geotargets.gdal.vector.driver",
-      "geotargets.gdal.vector.creation.options",
-      "geotargets.terra.preserve.metadata"
-    ))
+    rlang::arg_match0(
+      option_name,
+      c(
+        "geotargets.gdal.raster.driver",
+        "geotargets.gdal.raster.creation.options",
+        "geotargets.gdal.vector.driver",
+        "geotargets.gdal.vector.creation.options",
[…]
