+++ DESCRIPTION
@@ -47,7 +47,8 @@ Suggests:
     sf,
     stars,
     testthat (>= 3.0.0),
-    fs
+    fs,
+    spelling
 Config/testthat/edition: 3
 URL: https://github.com/njtierney/geotargets, http://geotargets.njtierney.com
 BugReports: https://github.com/njtierney/geotargets/issues
