### 2025-03-12 bfb39e0
first pass at converting URLs etc to ropensci
