+++ R/tar-terra-rast.R
@@ -34,7 +34,7 @@
 #'   slow pipelines. Also note metadata may be impacted by different versions
 #'   of GDAL and different drivers. If you have an issue with retaining
 #'   metadata for your setup, please file an issue at
-#'   \url{https://github.com/njtierney/geotargets/issues/} and we will try and
+#'   \url{https://github.com/ropensci/geotargets/issues/} and we will try and
 #'   get this working for you. Also note that you can specify this option
 #'   inside [geotargets_option_set()] if you want to set this for the entire
 #'   pipeline.
@@ -69,27 +69,27 @@
 #'   })
 #' }
 tar_terra_rast <- function(
-        name,
-        command,
-        pattern = NULL,
-        filetype = geotargets_option_get("gdal.raster.driver"),
-        gdal = geotargets_option_get("gdal.raster.creation.options"),
-        preserve_metadata = geotargets_option_get("terra.preserve.metadata"),
-        ...,
-        tidy_eval = targets::tar_option_get("tidy_eval"),
-        packages = targets::tar_option_get("packages"),
-        library = targets::tar_option_get("library"),
-        repository = targets::tar_option_get("repository"),
-        error = targets::tar_option_get("error"),
-        memory = targets::tar_option_get("memory"),
-        garbage_collection = targets::tar_option_get("garbage_collection"),
-        deployment = targets::tar_option_get("deployment"),
-        priority = targets::tar_option_get("priority"),
-        resources = targets::tar_option_get("resources"),
-        storage = targets::tar_option_get("storage"),
-        retrieval = targets::tar_option_get("retrieval"),
-        cue = targets::tar_option_get("cue"),
-        description = targets::tar_option_get("description")
+  name,
+  command,
+  pattern = NULL,
+  filetype = geotargets_option_get("gdal.raster.driver"),
+  gdal = geotargets_option_get("gdal.raster.creation.options"),
+  preserve_metadata = geotargets_option_get("terra.preserve.metadata"),
+  ...,
+  tidy_eval = targets::tar_option_get("tidy_eval"),
+  packages = targets::tar_option_get("packages"),
+  library = targets::tar_option_get("library"),
+  repository = targets::tar_option_get("repository"),
+  error = targets::tar_option_get("error"),
+  memory = targets::tar_option_get("memory"),
+  garbage_collection = targets::tar_option_get("garbage_collection"),
+  deployment = targets::tar_option_get("deployment"),
+  priority = targets::tar_option_get("priority"),
+  resources = targets::tar_option_get("resources"),
+  storage = targets::tar_option_get("storage"),
+  retrieval = targets::tar_option_get("retrieval"),
+  cue = targets::tar_option_get("cue"),
+  description = targets::tar_option_get("description")
 ) {
   filetype <- filetype %||% "GTiff"
 
@@ -129,17 +129,17 @@ tar_terra_rast <- function(
     format = targets::tar_format(
       read = tar_rast_read(preserve_metadata = preserve_metadata),
       write = tar_rast_write(
-          filetype = filetype,
-          gdal = gdal,
-          preserve_metadata = preserve_metadata
-          ),
+        filetype = filetype,
+        gdal = gdal,
+        preserve_metadata = preserve_metadata
+      ),
       marshal = function(object) terra::wrap(object),
       unmarshal = function(object) terra::unwrap(object),
       substitute = list(
-          filetype = filetype,
-          gdal = gdal,
-          preserve_metadata = preserve_metadata
-          )
+        filetype = filetype,
+        gdal = gdal,
+        preserve_metadata = preserve_metadata
+      )
     ),
     repository = repository,
[…]
