+++ man/geotargets-options.Rd
@@ -70,7 +70,7 @@ this could be due to the versions of GDAL and terra on your machine. We
 recommend exploring if this issue persists outside of geotargets. That is,
 try saving the file out and reading it back in using regular R code. If you
 find that this is an issue with geotargets, please file an issues at
-\url{https://github.com/njtierney/geotargets/issues/} and we will try and
+\url{https://github.com/ropensci/geotargets/issues/} and we will try and
 get this working for you.
 }
 
