+++ .github/workflows/test-coverage.yaml
@@ -46,4 +46,4 @@ jobs:
         uses: codecov/codecov-action@v4.0.1
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
-          slug: njtierney/geotargets
+          slug: ropensci/geotargets
