### 2025-03-12 cbbc40d
update more links to ropensci in README, DESCRIPTION, and also version bump and touch up NEWS.md
