+++ README.Rmd
@@ -13,7 +13,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# geotargets <a href="http://geotargets.njtierney.com/"><img src="man/figures/logo.png" alt="geotargets website" align="right" height="139"/></a> 
+# geotargets <a href="https://docs.ropensci.org/geotargets"><img src="man/figures/logo.png" alt="geotargets website" align="right" height="139"/></a> 
 
 <!-- badges: start -->
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -36,7 +36,7 @@ One example citation of geotargets could be as follows: "R packages used in this
   (2024). "geotargets: 'Targets'
   Extensions for Geospatial
   Formats."
-  <http://geotargets.njtierney.com>.
+  <https://docs.ropensci.org/geotargets>.
 
 ## Installation
 
