+++ NEWS.md
@@ -1,8 +1,9 @@
-# geotargets (development version)
+# geotargets 0.3.0 (12 March 2025)
 
 * bugfix by @brownag that fixes use of `file.rename()` in `tar_terra_rast(..., preserve_metadata = "zip")`, which does not work when the temporary directory is on a different partition. (#121, PR #122).
 * Fixed examples for `tar_terra_tiles()`, `tile_grid()`, `tar_terra_sds()`, and `tar_terra_sprc()` as reported by @amart90 as part of [rOpenSci review](https://github.com/ropensci/software-review/issues/675)
 * Added details to the documentation for `tar_terra_tiles()` (suggested by @amart90 as part of [rOpenSci review](https://github.com/ropensci/software-review/issues/675))
+* completed ropensci review and transferred ownership to ropensci
 
 # geotargets 0.2.0 (29 November 2024)
 
