### 2025-03-12 #675 @mpadge
Congrats Nick, awesome work! I'll answer that last question for you: Yes on both counts. Submit as normal, and this review will be taken into account there (so make sure you link back to here), and the whole process should be very quick.

Your other questions seem like they've also been auto-answered: https://docs.ropensci.org/geotargets/ is up and running.
