### 2025-03-12 #675 @njtierney
Thanks, Mark! OK great, I'll get a JOSS paper together and try and submit soon.

That's great to know about the ropensci docs working!
